phones/requirements.txt
python-socketio[client]==5.2.1
requests==2.25.1
psutil==5.8.0